# Conflicts:
#	common/pom.xml
#	common/src/main/java/io/bisq/common/app/Version.java
#	common/src/main/resources/i18n/displayStrings.properties
#	common/src/main/resources/i18n/displayStrings_hu.properties
#	common/src/main/resources/i18n/displayStrings_ro.properties
#	consensus/pom.xml
#	core/pom.xml
#	core/src/main/java/io/bisq/core/offer/OpenOfferManager.java
#	core/src/main/java/io/bisq/core/payment/PaymentAccountUtil.java
#	core/src/main/java/io/bisq/core/provider/fee/FeeService.java
#	gui/pom.xml
#	gui/src/main/java/io/bisq/gui/bisq.css
#	gui/src/main/java/io/bisq/gui/main/offer/createoffer/CreateOfferDataModel.java
#	gui/src/main/java/io/bisq/gui/main/offer/createoffer/CreateOfferView.java
#	gui/src/main/java/io/bisq/gui/main/offer/takeoffer/TakeOfferView.java
#	gui/src/main/java/io/bisq/gui/main/portfolio/openoffer/OpenOffersView.java
#	gui/src/main/java/io/bisq/gui/main/portfolio/openoffer/OpenOffersViewModel.java
#	gui/src/main/java/io/bisq/gui/util/GUIUtil.java
#	monitor/pom.xml
#	network/pom.xml
#	network/src/main/resources/PersistableNetworkPayloadMap_BTC_MAINNET
#	package/linux/32bitBuild.sh
#	package/linux/64bitBuild.sh
#	package/linux/Dockerfile
#	package/osx/create_app.sh
#	package/osx/finalize.sh
#	package/windows/32bitBuild.bat
#	package/windows/64bitBuild.bat
#	package/windows/Bisq.iss
#	pom.xml
#	seednode/pom.xml
#	statistics/pom.xml

+++ core/src/main/java/io/bisq/core/offer/OpenOfferManager.java
@@ -47,6 +47,7 @@ import io.bisq.network.p2p.*;
 import io.bisq.network.p2p.peers.PeerManager;
 import javafx.collections.ObservableList;
 import org.bitcoinj.core.Coin;
+import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -353,17 +354,26 @@ public class OpenOfferManager implements PeerManager.Listener, DecryptedDirectMe
 
     public void removeOpenOffer(OpenOffer openOffer, ResultHandler resultHandler, ErrorMessageHandler errorMessageHandler) {
         Offer offer = openOffer.getOffer();
-        offerBookService.removeOffer(offer.getOfferPayload(),
-                () -> {
-                    offer.setState(Offer.State.REMOVED);
-                    openOffer.setState(OpenOffer.State.CANCELED);
-                    openOffers.remove(openOffer);
-                    closedTradableManager.add(openOffer);
-                    log.debug("removeOpenOffer, offerId={}", offer.getId());
-                    btcWalletService.resetAddressEntriesForOpenOffer(offer.getId());
-                    resultHandler.handleResult();
-                },
-                errorMessageHandler);
+        if (openOffer.isDeactivated()) {
+            openOffer.setStorage(openOfferTradableListStorage);
+            onRemoved(openOffer, resultHandler, offer);
+        } else {
+            offerBookService.removeOffer(offer.getOfferPayload(),
+                    () -> {
+                        onRemoved(openOffer, resultHandler, offer);
+                    },
+                    errorMessageHandler);
+        }
+    }
+
+    private void onRemoved(@NotNull OpenOffer openOffer, ResultHandler resultHandler, Offer offer) {
+        offer.setState(Offer.State.REMOVED);
+        openOffer.setState(OpenOffer.State.CANCELED);
+        openOffers.remove(openOffer);
+        closedTradableManager.add(openOffer);
+        log.debug("removeOpenOffer, offerId={}", offer.getId());
+        btcWalletService.resetAddressEntriesForOpenOffer(offer.getId());
+        resultHandler.handleResult();
     }
 
     // Close openOffer after deposit published
